rename skypy.galaxy to skypy.galaxies
