+++ README.rst
@@ -7,7 +7,7 @@ SkyPy: A package for modelling the Universe
 This package contains methods for modelling the Universe, galaxies and the Milky
 Way. Also included are methods for generating observed data.
 
-* Galaxy_ morphology, luminosity and redshift distributions
+* Galaxies_: morphology, luminosity and redshift distributions
 * Pipelines_ to generate populations of astronomical objects
 
 The full list of features can be found in the `SkyPy Documentation`_.
@@ -15,7 +15,7 @@ The full list of features can be found in the `SkyPy Documentation`_.
 If you use SkyPy for work or research presented in a publication please follow
 our `Citation Guidelines`_.
 
-.. _Galaxy: https://skypy.readthedocs.io/en/latest/galaxy.html
+.. _Galaxies: https://skypy.readthedocs.io/en/latest/galaxies.html
 .. _Pipelines: https://skypy.readthedocs.io/en/latest/pipeline/index.html
 .. _SkyPy Documentation: https://skypy.readthedocs.io/en/latest/
 .. _Citation Guidelines: CITATION.rst
