+++ docs/pipeline/examples/config.yml
@@ -7,9 +7,9 @@ cosmology: !astropy.cosmology.FlatLambdaCDM
 n_galaxies: 100
 tables:
   galaxies:
-    redshift: !skypy.galaxy.redshift.smail
+    redshift: !skypy.galaxies.redshift.smail
       z_median: 1.0
       alpha: 1.5
       beta: 2.0
       size: $n_galaxies
-    distance_modulus: !skypy.galaxy.luminosity.distance_modulus [$galaxies.redshift]
+    distance_modulus: !skypy.galaxies.luminosity.distance_modulus [$galaxies.redshift]
