+++ docs/pipeline/examples/pipeline.yml
@@ -1,6 +1,6 @@
 parameters:
   median-redshift: 1.0
-galaxy-redshifts: !skypy.galaxy.redshift.smail
+galaxy-redshifts: !skypy.galaxies.redshift.smail
   z_median: $median-redshift
   alpha: 1.5
   beta: 2.0
