+++ examples/galaxies/sdss_photometry.yml
@@ -8,45 +8,45 @@ magnitude_limit: 23
 sky_area: 10 deg2
 tables:
   blue_galaxies:
-    redshift, magnitude: !skypy.galaxy.schechter_lf
+    redshift, magnitude: !skypy.galaxies.schechter_lf
       redshift: $z_range
       M_star: $M_star
       phi_star: $phi_star
       alpha: -1.3
       m_lim: $magnitude_limit
       sky_area: $sky_area
-    spectral_coefficients: !skypy.galaxy.spectrum.dirichlet_coefficients
+    spectral_coefficients: !skypy.galaxies.spectrum.dirichlet_coefficients
       alpha0: [2.079, 3.524, 1.917, 1.992, 2.536]
       alpha1: [2.265, 3.862, 1.921, 1.685, 2.480]
       weight: [3.47e+09, 3.31e+06, 2.13e+09, 1.64e+10, 1.01e+09]
       redshift: $blue_galaxies.redshift
-    stellar_mass: !skypy.galaxy.spectrum.kcorrect.stellar_mass
+    stellar_mass: !skypy.galaxies.spectrum.kcorrect.stellar_mass
       coefficients: $blue_galaxies.spectral_coefficients
       magnitudes: $blue_galaxies.magnitude
       filter: bessell-B
-    mag_u, mag_g, mag_r, mag_i, mag_z: !skypy.galaxy.spectrum.kcorrect.apparent_magnitudes
+    mag_u, mag_g, mag_r, mag_i, mag_z: !skypy.galaxies.spectrum.kcorrect.apparent_magnitudes
       coefficients: $blue_galaxies.spectral_coefficients
       redshift: $blue_galaxies.redshift
       filters: sdss2010-*
       stellar_mass: $blue_galaxies.stellar_mass
   red_galaxies:
-    redshift, magnitude: !skypy.galaxy.schechter_lf
+    redshift, magnitude: !skypy.galaxies.schechter_lf
       redshift: $z_range
       M_star: $M_star_red
       phi_star: $phi_star_red
       alpha: -0.5
       m_lim: $magnitude_limit
       sky_area: $sky_area
-    spectral_coefficients: !skypy.galaxy.spectrum.dirichlet_coefficients
+    spectral_coefficients: !skypy.galaxies.spectrum.dirichlet_coefficients
       alpha0: [2.461, 2.358, 2.568, 2.268, 2.402]
       alpha1: [2.410, 2.340, 2.200, 2.540, 2.464]
       weight: [3.84e+09, 1.57e+06, 3.91e+08, 4.66e+10, 3.03e+07]
       redshift: $red_galaxies.redshift
-    stellar_mass: !skypy.galaxy.spectrum.kcorrect.stellar_mass
+    stellar_mass: !skypy.galaxies.spectrum.kcorrect.stellar_mass
       coefficients: $red_galaxies.spectral_coefficients
       magnitudes: $red_galaxies.magnitude
       filter: bessell-B
-    mag_u, mag_g, mag_r, mag_i, mag_z: !skypy.galaxy.spectrum.kcorrect.apparent_magnitudes
+    mag_u, mag_g, mag_r, mag_i, mag_z: !skypy.galaxies.spectrum.kcorrect.apparent_magnitudes
       coefficients: $red_galaxies.spectral_coefficients
       redshift: $red_galaxies.redshift
       filters: sdss2010-*
