+++ skypy/galaxies/stellar_mass.py
