+++ skypy/galaxies/tests/test_import.py
@@ -1,0 +1,7 @@
+def test_import():
+    import skypy.galaxies
+    import skypy.galaxies.luminosity
+    import skypy.galaxies.morphology
+    import skypy.galaxies.redshift
+    import skypy.galaxies.spectrum
+    import skypy.galaxies.stellar_mass
