--- skypy/galaxy/tests/test_import.py
@@ -1,7 +1,0 @@
-def test_import():
-    import skypy.galaxy
-    import skypy.galaxy.luminosity
-    import skypy.galaxy.morphology
-    import skypy.galaxy.redshift
-    import skypy.galaxy.spectrum
-    import skypy.galaxy.stellar_mass
